Move algorithm/configuration.py to base
