+++ external/anomaly/configs/base/draem/configuration.py
@@ -5,7 +5,7 @@
 
 
 from attr import attrs
-from configs.base import BaseAnomalyConfig
+from configs.base.configuration import BaseAnomalyConfig
 from configs.base.configuration_enums import EarlyStoppingMetrics
 from ote_sdk.configuration.elements import (
     ParameterGroup,
