+++ external/anomaly/configs/base/padim/__init__.py
